pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
     "fastapi>=0.135.1",
     "motor>=3.7.1",
     "numpy>=2.4.3",
     "pandas>=3.0.1",
     "passlib[bcrypt]>=1.7.4",
     "python-dotenv>=1.2.2",
     "python-jose[cryptography]>=3.5.0",
     "python-multipart>=0.0.22",
     "slowapi>=0.1.9",
     "uvicorn>=0.41.0",
]
